Finished articles app and styling

+++ articles/views.py
@@ -1,9 +1,10 @@
-from django.shortcuts import render, get_object_or_404, redirect
+from django.shortcuts import render, get_object_or_404, redirect, reverse
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
 from .models import Article
 from .forms import ArticleForm
 
 
-# Create your views here.
 def article_list(request):
     articles = Article.objects.all()
     return render(request, 'articles/article_list.html', {'articles': articles})
@@ -14,32 +15,57 @@ def article_detail(request, pk):
     return render(request, 'articles/article_detail.html', {'article': article})
 
 
+@login_required
 def article_create(request):
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site admins can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ArticleForm(request.POST, request.FILES)
         if form.is_valid():
             form.save()
+            messages.success(request, 'Article created successfully.')
             return redirect('articles:article_list')
+        else:
+            messages.error(request, 'Failed to create the article. Please check the form.')
     else:
         form = ArticleForm()
-    return render(request, 'articles/article_form.html', {'form': form})
+    return render(request, 'articles/article_create.html', {'form': form, 'action': 'Create'})
 
 
+@login_required
 def article_update(request, pk):
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site admins can do that.')
+        return redirect(reverse('home'))
+
     article = get_object_or_404(Article, pk=pk)
     if request.method == 'POST':
         form = ArticleForm(request.POST, request.FILES, instance=article)
         if form.is_valid():
             form.save()
+            messages.success(request, 'Article updated successfully.')
             return redirect('articles:article_list')
+        else:
+            messages.error(request, 'Failed to update the article. Please check the form.')
     else:
         form = ArticleForm(instance=article)
-    return render(request, 'articles/article_form.html', {'form': form})
+    return render(request, 'articles/article_form.html', {'form': form, 'article': article})
 
 
+@login_required
 def article_delete(request, pk):
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site admins can do that.')
+        return redirect(reverse('home'))
+
     article = get_object_or_404(Article, pk=pk)
     if request.method == 'POST':
         article.delete()
+        messages.success(request, 'Article deleted successfully.')
         return redirect('articles:article_list')
     return render(request, 'articles/article_confirm_delete.html', {'article': article})
